changed app.js/index.js structure
