+++ client/components/Cart.js
@@ -66,7 +66,6 @@ class Cart extends Component {
 
     return userCart ? (
       <div>
-        <Navbar />
         <h3 className="allProductsTitle"> Shopping Cart</h3> <hr />
         <div className="some-page-wrapper">
           {userCart.map(product => {
@@ -120,7 +119,6 @@ class Cart extends Component {
       <GuestCart />
     ) : (
       <div>
-        <Navbar />
         <h2 className="allProductsTitle">Your Shopping Cart Is Empty</h2>
         <div>
           <img className="emptyCart" src="http://tny.im/ky1" />
