+++ client/components/Checkout.js
@@ -59,7 +59,6 @@ class Checkout extends Component {
     const orderId = this.props.userCart[0].id
     return user.id ? (
       <div>
-        <Navbar />
         <div>
           <div className="shippingrow">
             <div>
